requirements.txt
torch
facenet_pytorch
torchvision
numpy
pillow
opencv-python
gdown